fix: Edit unsplash provider to support image urls returned by the api by correctly appending the format query params.
I found out that when using URLs which already contains query parameters, the provider will wrongly attach a '?' Before the applied operations and Unsplash would ignore the first passed operation.
In these cases, placing a & would be a better option.
I think another workaround would be to always place a & after ? like this: ?&
Here's an example of one such link.