feat: add data-nuxt-img and data-nuxt-pic attrs
This commit adds data attributes to usages of nuxt/image
and nuxt/picture. This will allow us to compare performance
results (in aggregate) for apps that use the components
against apps that use normal <img> or <picture> tags.
Currently, it isn't possible because there isn't a good way
to distinguish the two in HTTP Archive.

Doing such a perf comparison allows us to:
- ensure that these components actually improve performance
- catch and fix performance regressions quickly