update feature detection markers to match new standard
Feature detection was added in #747 to allow performance
analysis for nuxt/image uses in aggregate. Since that time,
Javascript-based feature markers have been standardized
in the W3C group around the User Timing API:
https://www.w3.org/TR/user-timing/#dfn-mark_feature_usage

This commit updates the feature tracking to use the new
standard, which is designed to be performance-neutral.
The previous mechanism (the data- attributes) will be
removed in a future PR, once we can confirm it works
as expected end-to-end.